Trigger and function for s3 and lambda
I created a bash script for uploading the aws lambda function and implementing the trigger, given that the s3 bucket and target bucket exists. The target should be named "examplename"-resized . In the case the lambda function does not exist on the given account, you might need to install/download something (included in the script). This omits the need to upload large files. I haven't found a solution to avoid this entirely, so feel free to add. I'll port this to go somewhere next week. The trigger works quite effective after the script is run: if png/jpg is uploaded to the source bin, a resised version is uploaded to the target. The functionality will of course change, this is just to test the trigger.